aider: Replace unordered list of building model examples with a select dropdown.
# Aider chat conversation:

USER: rewrite the building model examples div to be a dropdown

ASSISTANT: To make this change we need to modify `templates/index.html`:

1. Replace the existing unordered list of building model examples with a select dropdown.

Here is the *SEARCH/REPLACE* block:

templates/index.html
```html
<<<<<<< SEARCH
<h4>223P Building Model Examples</h4>
<div id="model-examples">
    <ul class="example-models">
    {% for model_name, model_file in models.items() %}
        <li><a href="javascript: void(0);" @click="handleURL('{{ model_file }}', true)">{{ model_name }}</a></li>
    {% endfor %}
    </ul>
</div>
=======
<h4>223P Building Model Examples</h4>
<div id="model-examples">
    <select class="example-models" @change="handleURL($event.target.value, true)">
        <option value=''><a href="javascript: void(0);">Select a model</a></option>
        {% for model_name, model_file in models.items() %}
            <option value='{{ model_file }}'><a href="javascript: void(0);">{{ model_name }}</a></option>
        {% endfor %}
    </select>
</div>
>>>>>>> REPLACE
```

Please note that this change assumes that the `handleURL` method in your Vue.js application can handle being called with the model file URL directly from the `@change` event of the select dropdown. If this is not the case, additional changes to `static/js/app.js` may be required.